flytter metoder for initiering av samværsavtale til utils-fil
